move_rotate

+++ cub3D/cub3d_utils.c
@@ -6,13 +6,13 @@
 /*   By: iwoo <iwoo@student.42seoul.kr>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/03/28 21:22:10 by iwoo              #+#    #+#             */
-/*   Updated: 2020/03/29 11:56:25 by iwoo             ###   ########.fr       */
+/*   Updated: 2020/04/01 15:11:12 by iwoo             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_cub3d.h"
 
-double get_rad(int num)
+double get_rad(double num)
 {
 	return (num * (M_PI / 180));
 }
